V0.0.1 - Initial version
Implemented Verida VIP-2 and tested
