--- VDA-DID-Registry/node-test/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "dapp",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "type": "module",
-  "scripts": {
-    "start": "node index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@alch/alchemy-web3": "^1.2.5",
-    "body-parser": "^1.19.2",
-    "cors": "^2.8.5",
-    "ethers": "^5.6.1",
-    "express": "^4.17.3",
-    "web3": "^1.7.0"
-  }
-}
